refactor(api): extract hooks + commands handlers from system.rs (#3749 4/N)
Extracts the webhook trigger endpoints (`/hooks/wake`, `/hooks/agent`)
and chat command catalog endpoints (`/commands`, `/commands/{name}`)
out of `routes/system.rs` into a dedicated
`routes/hooks_commands.rs` module — 5 functions, ~440 lines.

* Public route paths and JSON shapes are unchanged — `system.rs::router()`
  now `.merge(crate::routes::hooks_commands::router())` so callers see no
  diff.
* `validate_webhook_token` (private helper) moves with the handlers since
  it is only called by `webhook_wake` / `webhook_agent`.
* `mod.rs` registers the new module and glob-re-exports it for
  utoipa-macro path compatibility.

Supersedes #4458 (rebase against current main was non-mechanical:
the 3749 series's parallel extractions of approvals / pairing /
task_queue rewrote enough of system.rs that the original branch's
diff no longer applied cleanly).

Refs #3749
