+++ crates/librefang-api/src/routes/mod.rs
@@ -24,6 +24,7 @@ pub mod budget;
 pub mod channels;
 pub mod config;
 pub mod goals;
+pub mod hooks_commands;
 pub mod inbox;
 pub mod mcp_auth;
 pub mod media;
@@ -63,6 +64,7 @@ pub use budget::*;
 pub use channels::*;
 pub use config::*;
 pub use goals::*;
+pub use hooks_commands::*;
 pub use inbox::*;
 pub use mcp_auth::*;
 pub use media::*;
